styling examples (#117)
* :rainbow: styling examples

examples that represent of the level of customization that we need in
dash-table. The examples are implemented with HTML tables to
demonstrate the intended behaviour.

I’d like to see each of these examples implemented with the
`dash_table` syntax. We’ll use these examples as our `dash-table`
documentation

* :x: removing black and pylint

this keeps tripping us up and I don’t think it’s worth the pain right
now.