Fix 'child in a list should have a unique key prop'
- Each element of a react list should have a unique key prop. This
is not the case here, as the key of a cell is generated from its
column index. To have unique key, we should use the row and column
index.

- The ./demo folder, should use the development version of react
(at least when you are developing ), otherwise React will NOT
show errors/warnings. This is very confusing, as the ./demo folder
is used for development purposes. The only way to see the react
warning was to change the script tags to point to the development
bundle of react. We should point to the correct bundle (e.g, if
NODE_ENV is set up to DEVELOPMENT, load the DEV bundle, if NODE_ENV
is set up to PRODUCTION, load the PRODUCTION bundle )

- I still get this error in the demo, but not in my component. I
guess there are some places where the keys are still not unique.
I was not able to find them in the dev tools and did not dig
thru the code too much